integration-tests: factour out Flow, Job, FlowManager
