+++ tests/network-tests/src/Flow.ts
@@ -1,0 +1,5 @@
+import { Api } from './Api'
+import { QueryNodeApi } from './QueryNodeApi'
+
+export type FlowArgs = { api: Api; env: NodeJS.ProcessEnv; query: QueryNodeApi }
+export type Flow = (args: FlowArgs) => Promise<void>
