+++ tests/network-tests/src/flows/proposals/updateRuntime.ts
@@ -1,5 +1,5 @@
 import BN from 'bn.js'
-import { FlowArgs } from '../../Scenario'
+import { FlowArgs } from '../../Flow'
 import { BuyMembershipHappyCaseFixture } from '../../fixtures/membershipModule'
 import { UpdateRuntimeFixture } from '../../fixtures/proposalsModule'
 import { PaidTermId } from '@joystream/types/members'
